asset centre verified

+++ recsys/asset/centre.py
@@ -4,26 +4,28 @@
 # Project    : Recommender Systems Lab: Towards State-of-the-Art                                   #
 # Version    : 0.1.0                                                                               #
 # Python     : 3.10.8                                                                              #
-# Filename   : /recsys/persistence/centre.py                                                       #
+# Filename   : /recsys/asset/centre.py                                                             #
 # ------------------------------------------------------------------------------------------------ #
 # Author     : John James                                                                          #
 # Email      : john.james.ai.studio@gmail.com                                                      #
 # URL        : https://github.com/john-james-ai/recsys-lab                                         #
 # ------------------------------------------------------------------------------------------------ #
 # Created    : Sunday March 19th 2023 04:20:33 pm                                                  #
-# Modified   : Sunday March 19th 2023 06:53:03 pm                                                  #
+# Modified   : Monday March 20th 2023 02:25:18 am                                                  #
 # ------------------------------------------------------------------------------------------------ #
 # License    : MIT License                                                                         #
 # Copyright  : (c) 2023 John James                                                                 #
 # ================================================================================================ #
 """Datasource Centre Module"""
 import os
+import shutil
 
-from sqlalchemy import engine
+from sqlalchemy import text, exc
 import pandas as pd
 
 from recsys.services.io import IOService
-from recsys.persistence.base import AssetCentreABC, Asset
+from recsys.asset.base import AssetCentreABC, Asset
+from recsys.persistence.database import Database
 
 
 # ------------------------------------------------------------------------------------------------ #
@@ -32,16 +34,22 @@ class AssetCentre(AssetCentreABC):
 
     Each subclass maintains the persistence of a single asset type.
 
-    Note: Each subclass must override the constructor designating location and tablename.
+    Note: Each subclass must override the constructor designating directory and tablename.
 
     """
 
     def __init__(
-        self, cxn: engine.connection, location: str, tablename: str, io: IOService = IOService
+        self,
+        database: Database,
+        directory: str,
+        tablename: str,
+        io: IOService = IOService,
     ) -> None:
-        self._cxn = cxn
+        super().__init__()
+        self._engine = database.engine
+        self._cxn = database.engine.connect()
         self._io = io
-        self._location = location
+        self._directory = directory
         self._tablename = tablename
 
     def add(self, asset: Asset) -> None:
@@ -51,12 +59,13 @@ class AssetCentre(AssetCentreABC):
             asset (Asset): The asset instance.
         Raises: FileExistsError if file already exists.
         """
-        if self.exists(name=asset.name):
-            msg = (
-                f"An asset named {asset.name} already exists. Change the name or replace the asset."
-            )
-            self._logger.error(msg)
-            raise FileExistsError(msg)
+        try:
+            if self.exists(name=asset.name):
+                msg = f"An asset named {asset.name} already exists. Change the name or replace the asset."
+                self._logger.error(msg)
+                raise FileExistsError(msg)
+        except exc.OperationalError:  # If first add, database won't exist at time of existence check.
+            pass
 
         self._register(asset)
         self._save(asset)
@@ -97,10 +106,32 @@ class AssetCentre(AssetCentreABC):
         asset_df = df[df["name"] == name]
         return asset_df.shape[0] > 0
 
+    def show(self) -> None:
+        """Prints the registry to screen."""
+        df = self._get_registry()
+        df = df.drop(["index"], axis=1, errors="ignore")  # Kluge to dump extra index column.
+        print(df)
+
+    def reset(self) -> None:
+        """Resets the asset registry and PURGES the asset repository."""
+        self._reset_registry()
+        self._purge_assets()
+
     def _register(self, asset: Asset) -> None:
         """Adds an asset to the registry."""
-        df = asset.to_df()
-        df.to_sql(name=self._tablename, con=self._cxn, if_exists="append")
+        df = asset.id
+        self._logger.debug(df)
+        try:
+            df.to_sql(
+                name=self._tablename,
+                con=self._cxn,
+                if_exists="append",
+                index=False,
+            )
+        except Exception:  # pragma: no cover
+            msg = f"Exception when attempting to add asset {asset.name} to the registry in the {self._tablename} table."
+            self._logger.error(msg)
+            raise
 
     def _unregister(self, name: str) -> None:
         """Removes an asset from the registry by name.
@@ -108,23 +139,29 @@ class AssetCentre(AssetCentreABC):
         Args:
             name (str): Name of the asset.
         """
-        df = self._get_registry()
-        df = df[df["name"] != name]
-        self._put_registry(df=df)
+        query = text(f"DELETE FROM {self._tablename} WHERE name='{name}';")
+        self._cxn.execute(query)
 
     def _get_registry(self) -> pd.DataFrame:
         """Returns the registry dataframe."""
-        return pd.read_sql("SELECT * FROM datasource")
+        try:
+            query = text(f"SELECT * from {self._tablename};")
+            return pd.read_sql(query, con=self._cxn)
 
-    def _put_registry(self, df: pd.DataFrame) -> None:
-        """Saves the registry, overwriting the existing registry. Used when deleting assets.
+        except Exception:  # pragma: no cover
+            msg = (
+                f"Exception when attempting to read the registry from the {self._tablename} table."
+            )
+            self._logger.error(msg)
+            raise
 
-        Note: This REPLACES the entire registry.
-
-        Args:
-            df (pd.DataFrame): The registry in DataFrame format.
-        """
-        df.to_sql(name=self._tablename, con=self._cxn, if_exists="replace")
+    def _reset_registry(self) -> None:
+        """Drops the registry table."""
+        query = text(f"DROP TABLE {self._tablename};")
+        try:
+            self._cxn.execute(query)
+        except exc.OperationalError:  # In case it doesn't exist.
+            pass
 
     def _save(self, asset: Asset) -> None:
         """Persists the asset to storage
@@ -134,7 +171,7 @@ class AssetCentre(AssetCentreABC):
 
         """
         filename = asset.name + ".pkl"
-        filepath = os.path.join(self._location, filename)
+        filepath = os.path.join(self._directory, filename)
         self._io.write(filepath=filepath, data=asset)
 
     def _load(self, name: str) -> Asset:
@@ -144,15 +181,20 @@ class AssetCentre(AssetCentreABC):
             name (str): The name for the asset
         """
         filename = name + ".pkl"
-        filepath = os.path.join(self._location, filename)
+        filepath = os.path.join(self._directory, filename)
         return self._io.read(filepath=filepath)
 
     def _delete(self, name: str) -> Asset:
-        """Deletes the asset from storage.
+        """Deletes the asset from storage (if it exists).
 
         Args:
             name (str): The name for the asset
         """
         filename = name + ".pkl"
-        filepath = os.path.join(self._location, filename)
-        os.remove(filepath)
+        filepath = os.path.join(self._directory, filename)
+        if os.path.exists(filepath):
+            os.remove(filepath)
+
+    def _purge_assets(self) -> None:
+        """Delete the directory containing assets."""
+        shutil.rmtree(self._directory, ignore_errors=True)
